# LineEndingsCheck: improve performance