--- a/.pytool/Plugin/LineEndingCheck/Readme.md
+++ b/.pytool/Plugin/LineEndingCheck/Readme.md
@@ -19,9 +19,15 @@ The plugin can be configured to ignore certain files.
 ``` yaml
 "LineEndingCheck": {
     "IgnoreFiles": []
+    "IgnoreFilesWithNoExtension": False
 }
 ```
 
 ### IgnoreFiles
 
 An **optional** list of git ignore patterns relative to the package root used to exclude files from being checked.
+
+### IgnoreFilesWithNoExtension
+
+An **optional** value that, if True, will insert the gitignore rules necessary to have this check ignore files
+that do not contain a file extension. Neccessary for binary files and/or POSIX like executables.