--- a/BaseTools/BaseTools.ci.yaml
+++ b/BaseTools/BaseTools.ci.yaml
@@ -44,7 +44,7 @@
         "AuditOnly": True,
     },
     "LineEndingCheck": {
-        "IgnoreFiles": ['*.*']
+        "IgnoreFiles": ["*"],
     }
     # MU_CHANGE [END]
 }