## Description

LineEndingsCheck generally takes anywhere from 4 to 6 seconds to run, depending on the size of the package. This change reduces the time to run by ~53% as tested on packages in MU_BASECORE due to the following changes:

1. Reorganized filters to reduce the amount of list comprehensions performed to filter out files that do not need to be checked

2. Organized the order of checks in regards to filters most likely to catch a file, and the expensiveness of the filter itself.

| Pkg                  | Before (s) | After (s) | Improvement |
|----------------------|------------|-----------|-------------|
| MdeModulePkg         | 5.50       | 3.00      | 45%         |
| MdePkg               | 5.72       | 2.88      | 50%         |
| NetworkPkg           | 4.30       | 2.09      | 51%         |
| PcAtChipsetPkg       | 3.98       | 1.79      | 55%         |
| PolicyServicePkg     | 4.12       | 1.87      | 55%         |
| ShellPkg             | 4.40       | 1.85      | 58%         |
| StandaloneMmPkg      | 4.17       | 1.72      | 59%         |
| UefiCpuPkg           | 4.13       | 2.12      | 49%         |
| UnitTestFrameworkPkg | 3.94       | 1.79      | 55%         |

- [ ] Impacts functionality?
  - **Functionality** - Does the change ultimately impact how firmware functions?
  - Examples: Add a new library, publish a new PPI, update an algorithm, ...
- [ ] Impacts security?
  - **Security** - Does the change have a direct security impact on an application,
    flow, or firmware?
  - Examples: Crypto algorithm change, buffer overflow fix, parameter
    validation improvement, ...
- [x] Breaking change?
  - **Breaking change** - Will anyone consuming this change experience a break
    in build or boot behavior?
  - Examples: Add a new library class, move a module to a different repo, call
    a function in a new library class in a pre-existing module, ...
- [ ] Includes tests?
  - **Tests** - Does the change include any explicit test code?
  - Examples: Unit tests, integration tests, robot tests, ...
- [ ] Includes documentation?
  - **Documentation** - Does the change contain explicit documentation additions
    outside direct code modifications (and comments)?
  - Examples: Update readme file, add feature readme file, link to documentation
    on an a separate Web page, ...

## How This Was Tested

Ensured the same number of files had the line endings check performed after filtering for all packages in MU_BASECORE

## Integration Instructions

Previously, only files with extensions were checked for valid line endings. Now, all files are checked for valid line endings.

If your package begins to fail because a file with no extension has invalid line endings and you do **not** want to update the line endings, you will need to do the following integration steps:

1. Ensure edk2-pytool-library is >= v0.16.2
2. Set IgnoreFilesWithNoExtension: True in the package config file:

```json
"LineEndingCheck": {
    "IgnoreFilesWithNoExtension": "True"
}
```