@makubacki 2023-08-08
> The pipelines indicate the filtering is off from before.

If you haven't already, can you please verify the results on a few platform repos as well?