@Javagedes 2023-08-09
The previous glob for files in the package was `*.*`, which is any file with an extension. One of the updates to this PR will be to glob all files, which includes files without an extension.

To enable packages to continue to filter out files without an extension, the rules described below needs to be added to the "IgnoreFiles" setting for the LineEndingsCheck. However an update to edk2-pytool-library needed to be made to support these rules (https://github.com/tianocore/edk2-pytool-library/pull/384).

Rule:
* 
!*.*
!*/