@Javagedes 2023-08-09
This PR is held up by needing to merge https://github.com/tianocore/edk2-pytool-library/pull/384. Once that is in and a release is out, edk2-pytool-library can be updated as a part of this PR.